New Feature: added character counting to post/page titles
This is to help limit the length of post and page titles to not break
the design. The input turns YELLOW when you have 10 or less characters
until hitting the limit. The input turns RED when you've hit the limit.

To change the maximum length, add your own function to the
max_post_title_length filter.